+++ db/migrate/20240316193959_add_request_warning_to_datasets.rb
@@ -1,0 +1,5 @@
+class AddRequestWarningToDatasets < ActiveRecord::Migration[6.1]
+  def change
+    add_column :datasets, :request_warning, :text
+  end
+end